requirements.txt
# Flask Web框架
flask==3.0.2

# Gunicorn WSGI服务器（生产环境）
gunicorn==21.2.0

# Flask跨域支持
flask-cors==4.0.0

# 环境变量管理
python-dotenv==1.0.1

# MySQL数据库连接器
mysql-connector-python==8.2.0

# HTTP请求库（用于API调用）
requests==2.31.0

# 异步网络库（可选，Windows环境可能需跳过）
gevent==23.9.1

# 注意：在Windows环境中，gevent可能需要C编译器支持
# 如果安装失败，可以跳过gevent安装，使用以下命令：
# pip install flask==3.0.2 flask-cors==4.0.0 python-dotenv==1.0.1 mysql-connector-python==8.2.0 requests==2.31.0